Moved more methods to ReportConfiguration
